+++ applications/external/protopirate/protocols/kia_v3_v4.h
@@ -10,7 +10,7 @@ void* kia_protocol_decoder_v3_v4_alloc(SubGhzEnvironment* environment);
 void kia_protocol_decoder_v3_v4_free(void* context);
 void kia_protocol_decoder_v3_v4_reset(void* context);
 void kia_protocol_decoder_v3_v4_feed(void* context, bool level, uint32_t duration);
-uint8_t kia_protocol_decoder_v3_v4_get_hash_data(void* context);
+uint32_t kia_protocol_decoder_v3_v4_get_hash_data(void* context);
 SubGhzProtocolStatus kia_protocol_decoder_v3_v4_serialize(
     void* context,
     FlipperFormat* flipper_format,
