+++ applications/external/protopirate/protocols/subaru.h
@@ -18,7 +18,7 @@ void* subghz_protocol_decoder_subaru_alloc(SubGhzEnvironment* environment);
 void subghz_protocol_decoder_subaru_free(void* context);
 void subghz_protocol_decoder_subaru_reset(void* context);
 void subghz_protocol_decoder_subaru_feed(void* context, bool level, uint32_t duration);
-uint8_t subghz_protocol_decoder_subaru_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_subaru_get_hash_data(void* context);
 SubGhzProtocolStatus subghz_protocol_decoder_subaru_serialize(
     void* context,
     FlipperFormat* flipper_format,
